move dhis2 views into dhis2.views
